`Removed unused files and moved SortBy component to a shared location`
